Clean up examples and consolidate test directories
- Reduce examples from 5 to 2 essential ones:
  - examples/basic: Simple example showing core functionality
  - examples/advanced: Advanced example with factory-based services and dependency injection
- Remove redundant examples: auto-dependencies, best-syntax, default-handlers, improved-dx, robust-naming
- Consolidate scattered test directories into tests/:
  - tests/factory-safety: Factory safety tests
  - tests/interface-conflicts: Interface conflict resolution tests
- Update CI workflow to build only the essential examples
- Update benchmark README to reflect new directory structure
- All examples now build successfully and demonstrate key features
