+++ .github/workflows/ci.yml
@@ -68,11 +68,8 @@ jobs:
     - name: Build examples
       run: |
         go build ./cmd/example
-        go build ./examples/auto-dependencies
-        go build ./examples/best-syntax
-        go build ./examples/default-handlers
-        go build ./examples/improved-dx
-        go build ./examples/robust-naming
+        go build ./examples/basic
+        go build ./examples/advanced
     
     - name: Build for multiple platforms
       run: |
