+++ tests/factory-safety/main.go
