+++ tests/interface-conflicts/main.go
