+++ tests/interface-conflicts/pkg1/database.go
