+++ tests/interface-conflicts/pkg2/database.go
